Upstream auth constants into shared/schema/Constants
