+++ services/app/src/components/views/Settings.tsx
@@ -1,7 +1,8 @@
 import Button from '@material-ui/core/Button';
 import {getContentSets, numPlayers} from 'app/actions/Settings';
 import * as React from 'react';
-import {CONTENT_SET_FULL_NAMES, URLS, VERSION} from '../../Constants';
+import {VERSION} from 'shared/schema/Constants';
+import {CONTENT_SET_FULL_NAMES, URLS} from '../../Constants';
 import {openWindow} from '../../Globals';
 import {DifficultyType, FontSizeType, MultiplayerState, SettingsType} from '../../reducers/StateTypes';
 import Card from '../base/Card';
