+++ services/quests/src/actions/User.tsx
@@ -3,8 +3,8 @@ import Redux from 'redux';
 import {registerUserAndIdToken} from 'shared/auth/API';
 import {loggedOutUser, UserState} from 'shared/auth/UserState';
 import {loginWeb as loginWebBase} from 'shared/auth/Web';
+import {AUTH_SETTINGS} from 'shared/schema/Constants';
 import {realtimeUtils} from '../Auth';
-import {API_HOST} from '../Constants';
 import {SetProfileMetaAction} from './ActionTypes';
 import {loadQuestFromURL} from './Quest';
 import {setSnackbar} from './Snackbar';
@@ -14,15 +14,6 @@ declare var window: any;
 export function setProfileMeta(user: UserState): SetProfileMetaAction {
   return {type: 'SET_PROFILE_META', user};
 }
-
-export const AUTH_SETTINGS = {
-  // Android: '545484140970-qrhcn069bbvae1mub2237h5k32mnp04k.apps.googleusercontent.com',
-  // iOS: (REVERSE_CLIENT_ID) '545484140970-lgcbm3df469kscbngg2iof57muj3p588.apps.googleusercontent.com',
-  API_KEY: 'AIzaSyCgvf8qiaVoPE-F6ZGqX6LzukBftZ6fJr8',
-  CLIENT_ID: (process && process.env && process.env.OAUTH2_CLIENT_ID) || '545484140970-jq9jp7gdqdugil9qoapuualmkupigpdl.apps.googleusercontent.com',
-  SCOPES: 'profile email',
-  URL_BASE: API_HOST,
-};
 
 export function loginUser(showPrompt: boolean, quest?: boolean | string): ((dispatch: Redux.Dispatch<any>) => void) {
   return (dispatch: Redux.Dispatch<any>) => {
